requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
pyyaml>=6.0
scikit-learn>=1.0.0
jupyter>=1.0.0
notebook>=6.4.0